add setup.py and move b*.py scripts to cli.py
